requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.55.2
pillow>=10.0.0
olmocr>=0.4.6
pytest>=7.0.0
pytest-mock>=3.10.0


